+++ JoyboyCommunity/src/components/avatar/props.ts
@@ -1,0 +1,4 @@
+export type AvatarProps = {
+  userId: string;
+  source?: string;
+};
