+++ JoyboyCommunity/src/components/divider/Divider.tsx
@@ -1,5 +1,4 @@
-import { View, Text } from "react-native";
-import React from "react";
+import { View } from "react-native";
 import styled from "styled-components/native";
 
 const Divider = styled(View)`
