+++ JoyboyCommunity/src/components/note/PostPage.tsx
@@ -1,16 +1,16 @@
 // src/screens/FeedScreen.js
 
-import React from "react";
-import { View, Text, Button, StyleSheet, TouchableOpacity, Platform } from "react-native";
+import { View, Text } from "react-native";
 import { Avatar } from "../avatar";
+import styles from "./styles";
 
-function PostPage({ navigation, post }) {
+export const PostPage: React.FC<{ post: any }> = ({ post }) => {
   const { content, author, timestamp, source, pubkey } = post;
 
   return (
     <View style={styles.card}>
       <View style={styles.authorContainer}>
-        <Avatar navigation={navigation} source={source} userId={pubkey} />
+        <Avatar source={source} userId={pubkey} />
         <Text style={styles.author}>{author}</Text>
         <Text style={styles.timestamp}>{timestamp}</Text>
       </View>
@@ -25,44 +25,4 @@ function PostPage({ navigation, post }) {
       ></TouchableOpacity> */}
     </View>
   );
-}
-
-const styles = StyleSheet.create({
-  card: {
-    // backgroundColor: "#022b3a",
-    width: Platform.OS != "android" ? "100%" : 250,
-    borderRadius: 8,
-    padding: 16,
-    marginBottom: 16,
-    shadowColor: "#000",
-    shadowOffset: {
-      width: 0,
-      height: 2,
-    },
-    shadowOpacity: 0.2,
-    shadowRadius: 2,
-    elevation: 2,
-  },
-  contentBox: {
-    padding: 8,
-    marginBottom: 8,
-  },
-  content: {
-    fontSize: 16,
-  },
-  authorContainer: {
-    flexDirection: "row",
-    justifyContent: "space-between",
-    alignItems: "center",
-  },
-  author: {
-    fontSize: 14,
-    color: "#333",
-  },
-  timestamp: {
-    fontSize: 12,
-    color: "#666",
-  },
-});
-
-export default PostPage;
+};
