+++ JoyboyCommunity/src/components/profile/CreateNote/styles.ts
@@ -1,0 +1,20 @@
+import { StyleSheet } from "react-native";
+
+export default StyleSheet.create({
+  container: {
+    marginVertical: 10,
+    marginHorizontal: 20,
+    color: "white",
+  },
+  input: {
+    height: 40,
+    borderColor: "gray",
+    borderWidth: 1,
+    paddingHorizontal: 10,
+    marginBottom: 10,
+    color: "white",
+  },
+  text: {
+    color: "white",
+  },
+});
