+++ JoyboyCommunity/src/components/profile/EditProfile/styles.ts
@@ -1,0 +1,16 @@
+import { StyleSheet } from "react-native";
+
+export default StyleSheet.create({
+  container: {
+    marginVertical: 10,
+    marginHorizontal: 20,
+  },
+  input: {
+    height: 40,
+    borderColor: "gray",
+    borderWidth: 1,
+    paddingHorizontal: 10,
+    marginBottom: 10,
+    color: "white",
+  },
+});
