+++ JoyboyCommunity/src/components/skeleton/LayoutContainer.tsx
@@ -1,4 +1,4 @@
-import { Platform, Text, View } from "react-native";
+import { Text, View } from "react-native";
 import { useSafeAreaInsets } from "react-native-safe-area-context";
 import { styled } from "styled-components/native";
 
@@ -14,8 +14,7 @@ const LayoutContainer = ({ title, children }) => {
     <Container style={{ flex: 1 }}>
       <View
         style={{
-    width: Platform.OS != "android" ? "100%" : 250,
-
+          width: "100%",
           padding: 16,
           paddingTop: insets.top,
         }}
@@ -24,6 +23,7 @@ const LayoutContainer = ({ title, children }) => {
           {title}
         </Text>
       </View>
+
       {children}
     </Container>
   );
