+++ JoyboyCommunity/src/components/skeleton/ScreenContainer.tsx
@@ -1,4 +1,4 @@
-import { SafeAreaView, View } from "react-native";
+import { SafeAreaView } from "react-native";
 import styled from "styled-components/native";
 
 const ScreenContainer = styled(SafeAreaView)`
