+++ JoyboyCommunity/src/types/post.ts
@@ -1,0 +1,9 @@
+export type Post = {
+  content: string;
+  author: string;
+  timestamp?: number;
+  source?: string;
+  id: string;
+  created_at: string;
+  pubkey: string;
+};
